Refactor import paths and enhance module organization
- Updated import paths across multiple modules to reflect the new structure under `utils/system/core`, ensuring consistent access to global variables and application control functionalities.
- Moved `global_vars` and application control functions into the `system/core` directory for better organization and maintainability.
- Adjusted related UI and database modules to utilize the updated import paths, enhancing code clarity and consistency.
- Enhanced the Python import helper with updated paths for improved code completion related to global variables and application control.
