+++ main.py
@@ -30,10 +30,10 @@ import sys
 import logging
 import matplotlib
 
-from utils import global_vars
-from utils.app_control import init_settings
+from utils.system.core import global_vars
+from utils.system.core.app_control import init_settings
 from utils.system.config.logging_config import setup_logger
-from utils.app_initialization import parse_arguments, initialize_app, setup_initial_app_state
+from utils.system.core.app_initialization import parse_arguments, initialize_app, setup_initial_app_state
 from utils.database.database import create_database
 from utils.robot.robot_control import update_database_from_usb
 from utils.ui.ui_setup import (initialize_main_window, setup_input_validation, connect_signal_handlers,
