+++ ui_files/BlinkingLabel.py
@@ -3,7 +3,7 @@ from PySide6.QtCore import QTimer, QRect, Qt
 from PySide6.QtGui import QFont
 from typing import Optional
 from utils.message.message import MessageType
-from utils import global_vars
+from utils.system.core import global_vars
 logger = global_vars.logger
 
 class BlinkingLabel(QLabel):
