+++ utils/__init__.py
@@ -1,6 +1,7 @@
 # utils_folder/__init__.py
 
-from .system.config.settings import Settings
+from .system.core import global_vars
 from .ui import notification_popup
+from .system.config.settings import Settings
 
-__all__ = ["UR10_Server_functions", "global_vars", "Settings"]
+__all__ = ['global_vars', 'notification_popup', 'Settings']
