+++ utils/server/UR10_Server_functions.py
@@ -1,4 +1,4 @@
-from utils import global_vars
+from utils.system.core import global_vars
 from PySide6.QtGui import QPixmap
 from utils.system.config.logging_config import setup_server_logger
