+++ utils/server/UR_Common_functions.py
@@ -5,7 +5,7 @@
 from typing import Literal, List, Optional, Union, Tuple
 
 from utils.database.database import load_from_database
-from utils import global_vars
+from utils.system.core import global_vars
 
 from utils.system.config.logging_config import setup_server_logger
