+++ utils/system/core/app_control.py
@@ -6,7 +6,7 @@ import os
 from PySide6.QtWidgets import QMessageBox
 from PySide6.QtCore import Qt
 
-from utils import global_vars
+from utils.system.core import global_vars
 from utils.ui.ui_helpers import set_settings_line_edits
 from utils.server.server import server_stop
 from utils.system.config.settings import Settings
